feat: 종류별 Gold — 사망 순간 판의 합계에 더하고 판이 끝나면 결산으로 진행 상태에 넣는다 (BC-002)
- 적 수치(EnemyStats)에 Gold를 둔다. 종류 에셋의 칸이고, 판 조립 때 적 수치 표에 고정된다.
  샘플 값은 [임시]: asteroid 1, moon 2, planet 5, star 10, comet 2.
- 사망 확정 절차(EnemyRoster.RecordDeath)가 그 적의 Gold를 이 판의 합계(World.EarnedGold)에 더한다.
  피해 사망과 파괴 요청 사망이 같고, 전투 정리로 치운 적은 더하지 않는다.
- 판은 진행 상태(PlayerState)를 바꾸지 않는다. 합계는 원자료(BattleRawData.EarnedGold)에 남고,
  오케스트레이터가 정리가 끝난 뒤 한 번 결산해 진행 상태에 더한다. 정리가 실패하면 결산하지 않는다.
  유저 데이터가 전투 밖에서만 바뀌므로 저장 시점도 전투 밖이다.
- 결산의 수령자는 참가자가 1명일 때 그 사람이다. 다인 플레이의 보상 귀속은 미정이다(d71c0f4의 전제와 같다).
- Gold를 int에서 long으로 바꾼다(원작 금액은 B·T 단위, 황금 배율 4200).
- 콘솔: 종류 설명에 Gold, 전투 시작·종료 콘솔에 진행 상태의 Gold와 진행 중인 판이 번 Gold, 원자료의 Gold.
- 계약 Enemy.DeathEarnsItsGoldForTheBattleAtOnce. CoreSmoke 41개 통과.

+++ Assets/BlackHole/Core/Players/PlayerState.cs
@@ -26,7 +26,8 @@ namespace BlackHole.Core
     public sealed class PlayerState
     {
         public PlayerId Id { get; }
-        public int Gold { get; private set; }
+        // 원작의 금액은 T(조) 단위까지 오르므로 int(약 21억)가 아니라 long이다.
+        public long Gold { get; private set; }
         // 진행 중인 전투에 들어가 있는가. 한 진행 상태는 한 번에 한 전투에만 들어간다.
         public bool InBattle { get; private set; }
 
@@ -35,8 +36,9 @@ namespace BlackHole.Core
             Id = id;
         }
 
-        // Gold를 더한다. 지금 부르는 곳은 없다 — 처치 보상이 붙으면 전투가 부른다.
-        public void EarnGold(int amount)
+        // Gold를 더한다. 전투 중에는 부르지 않는다 — 판이 끝난 뒤 결산이 그 판의 원자료(EarnedGold)로 한 번 부른다.
+        // 그래서 진행 상태는 전투 밖에서만 바뀌고, 저장 시점도 전투 밖이다.
+        public void EarnGold(long amount)
         {
             if (amount < 0)
                 throw new ArgumentOutOfRangeException(nameof(amount), "0 이상이어야 한다.");
